# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VA eBay Client Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── eBay Accounts ──────────────────────────────────────────────────────────
...
  /bank-accounts:
    get:
      operationId: getBankAccounts
      tags: [bank-accounts]
      summary: List all bank accounts
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [bank-accounts]
      summary: Create bank account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankAccountBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank-accounts/{id}:
    patch:
      operationId: updateBankAccount
      tags: [bank-accounts]
      summary: Update bank account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBankAccountBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
    delete:
      operationId: deleteBankAccount
      tags: [bank-accounts]
      summary: Delete bank account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Invoices ───────────────────────────────────────────────────────────────
...
  /earnings:
    get:
      operationId: getEarnings
      tags: [earnings]
      summary: List earnings records
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Earning"
    post:
      operationId: createEarning
      tags: [earnings]
      summary: Create earning record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEarningBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Earning"
  /earnings/{id}:
    patch:
      operationId: updateEarning
      tags: [earnings]
      summary: Update earning
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEarningBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Earning"
    delete:
      operationId: deleteEarning
      tags: [earnings]
      summary: Delete earning
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /daily-logins:
    get:
      operationId: getDailyLogins
      tags: [daily-login]
      summary: List daily login records
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyLogin"
    post:
      operationId: createDailyLogin
      tags: [daily-login]
      summary: Create or update daily login record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyLoginBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLogin"
  /daily-logins/{id}:
    patch:
      operationId: updateDailyLogin
      tags: [daily-login]
      summary: Update daily login
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDailyLoginBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLogin"
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        businessBrand:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        onboardedOn:
          type: ["string", "null"]
        status:
          type: string
        hourlyRate:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        ebayAccountsCount:
          type: integer
        totalSales:
          type: ["number", "null"]
        lastContact:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientName, status, ebayAccountsCount, createdAt, updatedAt]
    CreateClientBody:
      type: object
      properties:
        clientName:
          type: string
        businessBrand:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        onboardedOn:
          type: ["string", "null"]
        status:
          type: string
        hourlyRate:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        totalSales:
          type: ["number", "null"]
        lastContact:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [clientName, status]
    UpdateClientBody:
      type: object
      properties:
        clientName:
          type: ["string", "null"]
        businessBrand:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        onboardedOn:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        hourlyRate:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        totalSales:
          type: ["number", "null"]
        lastContact:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        bankCode:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        bankName:
          type: ["string", "null"]
        accountHolder:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        routingCode:
          type: ["string", "null"]
        swiftBic:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        onlineBankingUrl:
          type: ["string", "null"]
        onlineUsername:
          type: ["string", "null"]
        onlinePassword:
          type: ["string", "null"]
        twoFaMethod:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, bankCode, clientId, clientName, createdAt, updatedAt]
    CreateBankAccountBody:
      type: object
      properties:
        bankCode:
          type: string
        clientId:
          type: integer
        bankName:
          type: ["string", "null"]
        accountHolder:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        routingCode:
          type: ["string", "null"]
        swiftBic:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        onlineBankingUrl:
          type: ["string", "null"]
        onlineUsername:
          type: ["string", "null"]
        onlinePassword:
          type: ["string", "null"]
        twoFaMethod:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [bankCode, clientId]
    UpdateBankAccountBody:
      type: object
      properties:
        bankCode:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        bankName:
          type: ["string", "null"]
        accountHolder:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        routingCode:
          type: ["string", "null"]
        swiftBic:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        onlineBankingUrl:
          type: ["string", "null"]
        onlineUsername:
          type: ["string", "null"]
        onlinePassword:
          type: ["string", "null"]
        twoFaMethod:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Earning:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        year:
          type: integer
        month:
          type: integer
        amount:
          type: number
        currency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, clientName, year, month, amount, createdAt, updatedAt]
    CreateEarningBody:
      type: object
      properties:
        clientId:
          type: integer
        year:
          type: integer
        month:
          type: integer
        amount:
          type: number
        currency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [clientId, year, month, amount]
    UpdateEarningBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        year:
          type: ["integer", "null"]
        month:
          type: ["integer", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    DailyLogin:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        ebayUsername:
          type: string
        year:
          type: integer
        month:
          type: integer
        targetPerWeek:
          type: ["integer", "null"]
        loginDays:
          type: array
          items:
            type: integer
        doneThisMonth:
          type: integer
        adherencePct:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, clientName, ebayUsername, year, month, loginDays, doneThisMonth, adherencePct, createdAt, updatedAt]
    CreateDailyLoginBody:
      type: object
      properties:
        clientId:
          type: integer
        ebayUsername:
          type: string
        year:
          type: integer
        month:
          type: integer
        targetPerWeek:
          type: ["integer", "null"]
        loginDays:
          type: array
          items:
            type: integer
      required: [clientId, ebayUsername, year, month, loginDays]
    UpdateDailyLoginBody:
      type: object
      properties:
        targetPerWeek:
          type: ["integer", "null"]
        loginDays:
          type: array
          items:
            type: integer